Move tracking user meta to it's own class.
