+++ includes/User_Preferences.php
@@ -1,0 +1,67 @@
+<?php
+/**
+ * User Preferences class.
+ *
+ * Register user meta.
+ *
+ * @package   Google\Web_Stories
+ * @copyright 2020 Google LLC
+ * @license   https://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
+ * @link      https://github.com/google/web-stories-wp
+ */
+
+/**
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Google\Web_Stories;
+
+/**
+ * User Preferences class.
+ */
+class User_Preferences {
+	/**
+	 * Name of the user meta key used for opt-in.
+	 *
+	 * @var string
+	 */
+	const OPTIN_META_KEY = 'web_stories_tracking_optin';
+
+	/**
+	 * Initializes User_Preferences.
+	 *
+	 * Registers the setting in WordPress.
+	 *
+	 * @since 1.0.0
+	 *
+	 * @return void
+	 */
+	public function init() {
+		register_meta(
+			'user',
+			static::OPTIN_META_KEY,
+			[
+				'type'              => 'boolean',
+				'sanitize_callback' => 'rest_sanitize_boolean',
+				'default'           => false,
+				'show_in_rest'      => true,
+				'auth_callback'     => static function() {
+					return current_user_can( 'edit_user', get_current_user_id() );
+				},
+				'single'            => true,
+			]
+		);
+	}
+}
