+++ includes/uninstall.php
@@ -125,7 +125,7 @@ function delete_stories_post_meta() {
  * @return void
  */
 function delete_stories_user_meta() {
-	delete_metadata( 'user', 0, Tracking::OPTIN_META_KEY, '', true );
+	delete_metadata( 'user', 0, User_Preferences::OPTIN_META_KEY, '', true );
 }
 
 /**
